+++ docs/examples/bannercallout/placeAtTop.tsx
@@ -6,7 +6,7 @@ export default function Example() {
       <Flex alignItems="center" height="100%" justifyContent="center" width="100%">
         <Flex direction="column" gap={{ column: 4, row: 0 }}>
           <Flex alignItems="center" justifyContent="start">
-            <Icon accessibilityLabel="" color="error" icon="pinterest" size={32} />
+            <Icon accessibilityLabel="" color="brandPrimary" icon="pinterest" size={32} />
             <ButtonGroup>
               <Button color="transparent" iconEnd="arrow-down" text="Business" />
               <Button color="transparent" iconEnd="arrow-down" text="Create" />