+++ packages/gestalt/src/Masonry.tsx
@@ -118,12 +118,6 @@ type Props<T> = {
    * This is an experimental prop and may be removed in the future.
    */
   _logTwoColWhitespace?: (arg1: number) => void;
-  /**
-   * Temporal prop to sync gutter logic on full width layout refactor.
-   *
-   * This is an experimental prop and will be removed in the future.
-   */
-  _legacyFlexibleGutterLogic?: boolean;
 };
 
 type State<T> = {
@@ -497,7 +491,6 @@ export default class Masonry<
       scrollContainer,
       _twoColItems,
       _logTwoColWhitespace,
-      _legacyFlexibleGutterLogic,
     } = this.props;
     const { hasPendingMeasurements, measurementStore, width } = this.state;
     const { positionStore } = this;
@@ -514,7 +507,6 @@ export default class Masonry<
         width,
         logWhitespace: _logTwoColWhitespace,
         _twoColItems,
-        _legacyFlexibleGutterLogic,
       });
     } else if (layout === 'uniformRow') {
       getPositions = uniformRowLayout({