+++ packages/gestalt/src/MasonryV2.tsx
@@ -136,12 +136,6 @@ type Props<T> = {
    * Experimental prop to trigger rendering updates via requestAnimationFrame
    */
   _useRAF?: boolean;
-  /**
-   * Temporal prop to sync gutter logic on full width layout refactor.
-   *
-   * This is an experimental prop and will be removed in the future.
-   */
-  _legacyFlexibleGutterLogic?: boolean;
 };
 
 type MasonryRef = {
@@ -338,7 +332,6 @@ function useLayout<
   _logTwoColWhitespace,
   _measureAll,
   _useRAF,
-  _legacyFlexibleGutterLogic,
 }: {
   align: Align;
   columnWidth: number;
@@ -353,7 +346,6 @@ function useLayout<
   _logTwoColWhitespace?: (arg1: number) => void;
   _measureAll?: boolean;
   _useRAF?: boolean;
-  _legacyFlexibleGutterLogic?: boolean;
 }): {
   height: number;
   hasPendingMeasurements: boolean;
@@ -378,7 +370,6 @@ function useLayout<
     width,
     _twoColItems,
     _logTwoColWhitespace,
-    _legacyFlexibleGutterLogic,
   });
 
   const itemMeasurements = items.filter((item) => measurementStore.has(item));
@@ -419,6 +410,7 @@ function useLayout<
     // - items: if we get new items, we should always recalculate positions
     // - itemMeasurementsCount: if we have a change in the number of items we've measured, we should always recalculage
     // - canPerformLayout: if we don't have a width, we can't calculate positions yet. so recalculate once we're able to
+    // eslint-disable-next-line react-compiler/react-compiler
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [itemMeasurementsCount, items, canPerformLayout]);
 
@@ -605,7 +597,6 @@ function Masonry<
     _logTwoColWhitespace,
     _measureAll,
     _useRAF,
-    _legacyFlexibleGutterLogic,
   }: Props<T>,
   ref:
     | {
@@ -676,6 +667,7 @@ function Masonry<
     if (!hasSetInitialWidth.current && width != null) {
       hasSetInitialWidth.current = true;
     }
+    // eslint-disable-next-line react-compiler/react-compiler
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [width]);
 
@@ -695,7 +687,6 @@ function Masonry<
     _logTwoColWhitespace,
     _measureAll,
     _useRAF,
-    _legacyFlexibleGutterLogic,
   });
 
   useFetchOnScroll({