+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+## 153.9.0 (Jun 19, 2024)
+
+### Minor
+
+- Masonry: Remove legacy flexible gutter logic ([#3639](https://github.com/pinterest/gestalt/pull/3639)) - [Preview link](https://deploy-preview-3639--gestalt.netlify.app?devexample=true)
+
+## 153.8.0 (Jun 18, 2024)
+
+### Minor
+
+- BannerSlim: Standardize dismiss functionality ([#3633](https://github.com/pinterest/gestalt/pull/3633)) - [Preview link](https://deploy-preview-3633--gestalt.netlify.app?devexample=true)
+
+## 153.7.1 (Jun 17, 2024)
+
+### Patch
+
+- Docs: fix Flex page ([#3634](https://github.com/pinterest/gestalt/pull/3634)) - [Preview link](https://deploy-preview-3634--gestalt.netlify.app?devexample=true)
+
+## 153.7.0 (Jun 17, 2024)
+
+### Minor
+
+- Flex: Added forwardRef to Flex, Flex.Item ([#3626](https://github.com/pinterest/gestalt/pull/3626)) - [Preview link](https://deploy-preview-3626--gestalt.netlify.app?devexample=true)
+
+## 153.6.1 (Jun 14, 2024)
+
+### Patch
+
+- SearchGuide: docs on mobile (iOS/Android) ([#3629](https://github.com/pinterest/gestalt/pull/3629)) - [Preview link](https://deploy-preview-3629--gestalt.netlify.app?devexample=true)
+
+## 153.6.0 (Jun 14, 2024)
+
+### Minor
+
+- Tokens: Add line-height tokens ([#3623](https://github.com/pinterest/gestalt/pull/3623)) - [Preview link](https://deploy-preview-3623--gestalt.netlify.app?devexample=true)
+
+## 153.5.0 (Jun 14, 2024)
+
+### Minor
+
+- ButtonToggle: V2 Development ([#3611](https://github.com/pinterest/gestalt/pull/3611)) - [Preview link](https://deploy-preview-3611--gestalt.netlify.app?devexample=true)
+
+## 153.4.0 (Jun 13, 2024)
+
+### Minor
+
+- Tokens, Icon: enables beta VR theme experimentation ([#3616](https://github.com/pinterest/gestalt/pull/3616)) - [Preview link](https://deploy-preview-3616--gestalt.netlify.app?devexample=true)
+
+## 153.3.2 (Jun 13, 2024)
+
+### Patch
+
+- Internal: adding react-compiler plugin and rule ([#3627](https://github.com/pinterest/gestalt/pull/3627)) - [Preview link](https://deploy-preview-3627--gestalt.netlify.app?devexample=true)
+
 ## 153.3.1 (Jun 12, 2024)
 
 ### Patch